mybootloader & graphics

+++ kernel/include/drivers/graphics.h
@@ -4,16 +4,16 @@
 
 void graphics_init();
 
-void set_pixel(int x, int y, uint8_t color);
+void set_pixel(int x, int y, uint32_t color);
 
 /* Bresenham's Line Algorithm */
-void draw_line(int x0, int y0, int x1, int y1, uint8_t color);
+void draw_line(int x0, int y0, int x1, int y1, uint32_t color);
 
-void draw_rectangle(int x, int y, int width, int height, uint8_t color);
+void draw_rectangle(int x, int y, int width, int height, uint32_t color);
 
 /* Midpoint Circle Algorithm */
-void draw_circle(int xc, int yc, int radius, uint8_t color);
+void draw_circle(int xc, int yc, int radius, uint32_t color);
 
-void graphics_clear(uint8_t color);
+void graphics_clear(uint32_t color);
 
 void graphics_update();
